if phase2 is set to rebuild, there could be new things that need to be downloaded, force phase 1 to restart if this is the case, as there could be new things that need to be downloaded. phase 2 is meant to not rely on internet connectivity
git-svn-id: https://svn.code.sf.net/p/rebeccablackos/code@4776 b52b6941-3400-464c-9f42-43200397181c

+++ externalbuilders/rebeccablackos_phase1.sh
@@ -41,7 +41,10 @@ fi
 mount --make-rprivate /
 
 #copy the dselect data saved in phase 2 into phase 1
-cp "$BUILDLOCATION"/build/"$BUILDARCH"/phase_2/tmp/INSTALLSSTATUS.txt "$BUILDLOCATION"/build/"$BUILDARCH"/phase_1/tmp/INSTALLSSTATUS.txt
+if [[ -f "$BUILDLOCATION"/build/"$BUILDARCH"/phase_2/tmp/INSTALLSSTATUS.txt ]]
+then
+  cp "$BUILDLOCATION"/build/"$BUILDARCH"/phase_2/tmp/INSTALLSSTATUS.txt "$BUILDLOCATION"/build/"$BUILDARCH"/phase_1/tmp/INSTALLSSTATUS.txt
+fi
 
 if [[ $HASOVERLAYFS == 0 ]]
 then
